[ALPS05623692] cam: fix mux usage
extend the number of mux

MTK-Commit-Id: 5de6a10657b0fc6bde6ca8260bca223621c22755

CR-Id: ALPS05623692
Feature: [Android Default] Camera Application Basic Functions

+++ drivers/misc/mediatek/imgsensor/src-v4l2/adaptor-subdrv.h
@@ -70,7 +70,7 @@ struct subdrv_ctx {
 	u8 mirror; /* mirrorflip information */
 	u8 sensor_mode; /* record IMGSENSOR_MODE enum value */
 	u32 shutter; /* current shutter */
-	u16 gain; /* current gain */
+	u32 gain; /* current gain */
 	u32 pclk; /* current pclk */
 	u32 frame_length; /* current framelength */
 	u32 line_length; /* current linelength */
